Added new libraries
Tong modified the libraries to make them work with analyzer 1.6. Hanjie has also made FADC classes for Scintillators and Cherenkov. These have been added to the libraries directory and the rootlogon.C
When put on a machine, go into the library directories and run 'make realclean' followed by 'make'.
For the FADC classes in libraries/Tritium_Hangie_Scin/ two make files will need to be run. There is a Makefile_Scin and Makefile_Cherenkov. Copy each to 'Makefile' and run the commands. This will be cleaned up later but it works for now.