SRC Fall Raster Calibrations! :D
Yaaaaaaaay
The time at 4.3 GeV after raster 1 died should be good, but is
potentially not the "best". There were no carbon foil or multifoil runs
taken after the raster was re-set-up. Nor were there any hydrogen
elastic runs taken. I was able to use a dummy run to optimize the
horizontal raster, however that is not as optimal since the "endcap" is
thicker.